feat: 딜러, 겜블러의 수익금 입출력 로직 구현

+++ src/main/kotlin/blackjack/ui/BlackjackReader.kt
@@ -25,7 +25,7 @@ object BlackjackReader {
         return readLine == YES_SIGN
     }
 
-    fun readBetAmount(gamblerName: String): Int {
+    fun readBetAmount(gamblerName: String): Double {
         BlackjackPrinter.askBetAmount(gamblerName)
         return ConsoleReader.readNumber()
     }
